Fixes view count inconsistency and tweaks chart
View counts were inconsistent because of how Impressionist works. Including
'has_many :impressions' caused Impressionist to use the number of impressions
that a user has *made* instead of the number of impressions on the user when
asking for 'impressionist_count'. This has been fixed and 'impressions' now
refers to impressions on the object and not made by the object.

Chart tweaking was done because the round function does not strictly round up
and was cutting of views. For instance, 12 views showed up as 10, until 15 views
were reached. Now, a chart with a max of 12 views will span the y axis to 20 as
originally intended.